requirements.txt
Django==4.2.15
djangorestframework==3.14.0
drf-yasg==1.21.7
django-cors-headers==3.11.0
python-dotenv==1.0.1
psycopg==3.2.1
